app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pika
python-dotenv
apscheduler
pydantic
python-multipart
pandas
pytest
httpx
requests
 